[CL 4730169 by Marcus Wassmer in Dev-Rendering branch]

+++ Engine/Source/Runtime/ApplicationCore/Private/Windows/WindowsPlatformApplicationMisc.cpp
@@ -319,7 +319,7 @@ float FWindowsPlatformApplicationMisc::GetDPIScaleFactorAtPoint(float X, float Y
 	{
 		if (GetDpiForMonitor)
 		{
-			POINT Position = { X, Y };
+			POINT Position = { static_cast<LONG>(X), static_cast<LONG>(Y) };
 			HMONITOR Monitor = MonitorFromPoint(Position, MONITOR_DEFAULTTONEAREST);
 			if (Monitor)
 			{
